switch from o.completedAt to o.number
makes full sense since this find recent method is used for generating order numbers for completed orders rather. If we say otherwise, we will have occasions in which the list returns empty for an order with a number 00001 and yet it does not have a completedAt.

Basically it makes sense because of how it is being used.